README.md
The "Love Tester" project developed in Python is a modern and digital approach to a classic game that evaluates the love compatibility between two people. Using the tkinter library, the program offers a friendly graphical interface, allowing users to enter their names and receive a percentage that indicates the level of love compatibility between them.

Under the hood, the random library is employed to generate a random result, ensuring that each test is unique and unpredictable. Although the result is generated in a playful way and has no scientific basis, it provides moments of fun and laughter for users.